Drop `distro-info` from clients requirements
The distro-info package is a special package that has to be
maintained by the distribution and is as such not suitable
for use as a wheel. The version we consume here has unknown
origins and has not been updated on pypi since 2013. It also
fails to install with the new pip dependency resolver.

Also fix name of `websocket-client`.